+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+* Types: improve Alerts types
+* Addons: allow fetching addon providers with authentication
+
 ## 0.3.4
 
 * User preferences: add `logs_font_size` and `theme` fields